OptiX

Documentation for OptiX.

OptiX.OhlcvBar

OptiX.convertAlphaVantageTimeSeriesDaily

OptiX.convertAlphaVantageTimeSeriesDailyBar

OptiX.findFiles

OptiX.readGZippedJson

OptiX.OhlcvBar — Type

Internal representation of generic time series data

OptiX.convertAlphaVantageTimeSeriesDaily — Method

Convert all bars of an AlphaVantage TIMESERIESDAILY call into a Vector of bars of our generic internal form

OptiX.convertAlphaVantageTimeSeriesDailyBar — Method

Convert an AlphaVantage TIMESERIESDAILY bar into our generic internal form

OptiX.findFiles — Method

Find all relevant files in my data lake

OptiX.readGZippedJson — Method

Read the contents of a gzipped json file into a JSON3 Object