OptiX

Documentation for OptiX.

OptiX.OhlcvBar

Base.hash

Base.isequal

Base.isless

OptiX.accumulateOhlcvBars

OptiX.convertAlphaVantageTimeSeriesDaily

OptiX.convertAlphaVantageTimeSeriesDailyBar

OptiX.convertAlphaVantageTimeSeriesIntraday

OptiX.convertAlphaVantageTimeSeriesIntradayBar

OptiX.createTimestamps

OptiX.deserializeOhlcvBars

OptiX.findFiles

OptiX.mergeBars

OptiX.prettyStringOhlcvBars

OptiX.readGZippedJson

OptiX.serializeOhlcvBars

OptiX.synthesizeOhlcvBar

OptiX.synthesizeOhlcvBars

OptiX.OhlcvBar — Type

Internal representation of generic time series data

Base.hash — Method

hash function for OhlcvBar s

Base.isequal — Method

Equality predicate for OhlcvBar s

Base.isless — Method

Comparison predicate for sorting OhlcvBar s

OptiX.accumulateOhlcvBars — Method

Use deserializeOhlcvBars, serializeOhlcvBars, and mergeBars to create the ability to read a collection of gzipped JSON files and merge their contents into a single Vector{OhlcvBar} file

OptiX.convertAlphaVantageTimeSeriesDaily — Method

Convert all bars of an AlphaVantage TIMESERIESDAILY call into a Vector of bars of our generic internal form

OptiX.convertAlphaVantageTimeSeriesDailyBar — Method

Convert an AlphaVantage TIMESERIESDAILY bar into our generic internal form

OptiX.convertAlphaVantageTimeSeriesIntraday — Method

Convert all bars of an AlphaVantage TIMESERIESINTRADAY call into a Vector of bars of our generic internal form

OptiX.convertAlphaVantageTimeSeriesIntradayBar — Method

Convert an AlphaVantage TIMESERIESINTRADAY bar into our generic internal form

OptiX.createTimestamps — Method

Create a range of timestamps given a start, stop, and step (eg Dates.Day(1))

OptiX.deserializeOhlcvBars — Method

Deserialize file to Vector{OhlcvBar}

OptiX.findFiles — Method

Find all relevant files in my data lake

OptiX.mergeBars — Method

Merge two Vector{OhlcvBar}s into one sorted Vector{OhlcvBar} without duplicates

OptiX.prettyStringOhlcvBars — Method

Pretty print Vector{OhlcvBar} to String

OptiX.readGZippedJson — Method

Read the contents of a gzipped json file into a JSON3 Object or JSON3.Array, as appropriate

OptiX.serializeOhlcvBars — Method

Serialize Vector{OhlcvBar} to a file

OptiX.synthesizeOhlcvBar — Method

Synthesize an OhlcvBar using values hashed on timestamp

OptiX.synthesizeOhlcvBars — Method

Leverage synthesizeOhlcvBar to synthesize a Vector(OhlcvBar) given a range of timestamps